refactor: 에러 상태 단일값 cvc, password -> 변경
