+++ README.md.md
@@ -68,7 +68,7 @@
 추가 리팩토링 사항 1
 
 - [x] 훅 결합 (상태 훅 + 에러 훅 결합)
-- [ ] 에러 상태 단일값 cvc, password -> 변경
+- [x] 에러 상태 단일값 cvc, password -> 변경
 
 추후 반영
