+++ src/hooks/useCardInformation/usePassword.ts
@@ -2,25 +2,21 @@ import { useState } from "react";
 import { PasswordType, SetValueFn, UseHookReturn } from "../../types/CardInformationType";
 import { PASSWORD_MAX_LENGTH } from "../../constants/constant";
 import useCheckLengthComplete from "../common/useCheckLengthComplete";
-import { errorStateType } from "../../types";
 import { validateNumberOnly } from "../../utils/validation";
 import useErrorCheckComplete from "../common/useErrorCheckComplete";
+import { SingleErrorType } from "../../types";
 
 const usePassword = (): UseHookReturn<"password"> => {
   const [password, setPasswordState] = useState<PasswordType>("");
-  const [isError, setIsError] = useState<errorStateType>([false]);
+  const [isError, setIsError] = useState<SingleErrorType>(false);
   const [errorMessage, setErrorMessage] = useState("");
 
   const setPassword: SetValueFn<PasswordType> = (value) => setPasswordState(value);
 
-  const validateInput = (value: string, index: number) => {
+  const validateInput = (value: string) => {
     const { error, message } = validateNumberOnly(value);
 
-    setIsError((prev) => {
-      const updated = [...prev];
-      updated[index] = error;
-      return updated;
-    });
+    setIsError(error);
 
     setErrorMessage(message);
   };
