--- src/types/useValidationType.ts
@@ -1,21 +1,0 @@
-import { cardInformationKey, errorStateType } from ".";
-
-export type useEachValidationType = {
-  isError: errorStateType;
-  isComplete: boolean;
-  errorMessage: string;
-  validateInput: (index: number, value: string) => void;
-};
-
-export type useValidationType = {
-  validation: validationFieldType;
-  isErrorCompletes: isErrorCompletesType;
-};
-
-export type validationFieldType = {
-  [K in Exclude<cardInformationKey, "company">]: Exclude<useEachValidationType, "isComplete">;
-};
-
-export type isErrorCompletesType = {
-  [K in Exclude<cardInformationKey, "company">]: Extract<useEachValidationType, "isComplete">;
-};
